requirements.txt
fastapi
uvicorn
httpx
python-dotenv
torch>=2.0.0
transformers>=4.30.0
datasets
numpy
scikit-learn